完成批量删除权限信息功能

+++ tt-manager/tt-manager-pojo/src/main/java/com/lkf/ttshop/pojo/po/RppermissionExample.java
@@ -145,7 +145,7 @@ public class RppermissionExample {
             return (Criteria) this;
         }
 
-        public Criteria andIdIn(List<BigDecimal> values) {
+        public Criteria andIdIn(List<Long> values) {
             addCriterion("Id in", values, "id");
             return (Criteria) this;
         }
